+++ modules/marker/handle_periodic_operations.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"strings"
 
-	markertypes "github.com/MonikaCat/provenance/x/marker/types"
-	"github.com/forbole/bdjuno/v3/modules/pricefeed/coingecko"
-	"github.com/forbole/bdjuno/v3/modules/utils"
-	"github.com/forbole/bdjuno/v3/types"
+	markertypes "github.com/MonCatCat/provenance/x/marker/types"
+	"github.com/forbole/bdjuno/v4/modules/pricefeed/coingecko"
+	"github.com/forbole/bdjuno/v4/modules/utils"
+	"github.com/forbole/bdjuno/v4/types"
 	"github.com/go-co-op/gocron"
 	"github.com/rs/zerolog/log"
 )