+++ CHANGELOG.md
@@ -10,6 +10,9 @@
 #### Parse Command
 - ([\#492](https://github.com/forbole/bdjuno/pull/492)) Add parse command for periodic tasks: `x/bank` total supply, `x/distribution` community pool, `x/mint` inflation, `pricefeed` token price and price history, `x/staking` staking pool
 
+#### Provenance
+- ([\#534](https://github.com/forbole/bdjuno/pull/534)) Upgrade provenance version to `v1.14.1` and proposals parsing to correctly parse `v0.46.x` gov proposals
+
 #### Upgrade Module
 - ([\#467](https://github.com/forbole/bdjuno/pull/467)) Store software upgrade plan and refresh data at upgrade height